requirements.txt
# Basic dependencies
flask>=2.0.0  # Web service framework
fastapi>=0.68.0  # API framework, optional alternative to Flask
uvicorn>=0.15.0  # ASGI server for running FastAPI

# MCP protocol
mcp>=1.3.0  # MCP protocol SDK

# Vector database
chromadb>=0.4.0  # Vector database for storing and retrieving code embeddings

# Embedding models
sentence-transformers>=2.2.0  # For generating code embedding vectors
torch>=1.10.0  # PyTorch for running embedding models

# Code parsing
tree-sitter>=0.20.0  # Code parsing library for structured code understanding
pygments>=2.10.0  # Code highlighting and syntax analysis

# Utility libraries
numpy>=1.20.0  # Numerical computation
pandas>=1.3.0  # Data processing
tqdm>=4.62.0  # Progress bar display
pyyaml>=6.0  # YAML configuration file processing
python-dotenv>=0.19.0  # Environment variable management